succ

+++ app/domains/integration/models.py
@@ -1,5 +1,5 @@
 # app\domains\integration\models.py
-from sqlalchemy import Column, BigInteger, String, Enum, DateTime, Boolean, ForeignKey, Text, JSON, func
+from sqlalchemy import Column, Integer, String, Enum, DateTime, Boolean, ForeignKey, Text, JSON, func
 from app.infra.database.base import Base
 import enum
 
@@ -13,8 +13,8 @@ class ServiceType(str, enum.Enum):
 class Integration(Base):
     __tablename__ = "integrations"
 
-    id               = Column(BigInteger, primary_key=True, autoincrement=True)
-    workspace_id     = Column(BigInteger, ForeignKey("workspaces.id"), nullable=False)
+    id               = Column(Integer, primary_key=True, autoincrement=True)
+    workspace_id     = Column(Integer, ForeignKey("workspaces.id"), nullable=False)
     service          = Column(Enum(ServiceType), nullable=False)
     access_token     = Column(Text, nullable=True)
     refresh_token    = Column(Text, nullable=True)
